godi_ph_api/requirements.txt
Django==3.0.5
django-cors-headers==3.3.0
PyMySQL==0.9.3
djangorestframework==3.11.0
django_mysql==3.8.1
django-redis==4.12.1
django-crontab==0.7.1
influxdb==5.3.0
numpy==1.18.5
apscheduler==3.6.3
tzlocal==2.1